Add saturation effect for dog
When "adding" hunger to dog, if there is "excess" hunger then that
will be added to the saturation hunger instead
Saturation Hunger :
When larger than Zero
+every hunger the dog lost will count against it
instead of the real hunger value.
+ If health < maxHealth : The dog will heal 1 health each 0.5s interval
with the cost of 3 saturation hunger per health.
NOTICE : The dog hunger value still caps at
the real max value.

Add saturation effect for dog

When "adding" hunger to dog, if there is "excess" hunger then that
will be added to the saturation hunger instead
Saturation Hunger :
When larger than Zero
+every hunger the dog lost will count against it
instead of the real hunger value.
+ If health < maxHealth : The dog will heal 1 health each 0.5s interval
with the cost of 3 saturation hunger per health.
NOTICE : The dog hunger value still caps at
the real max value.