Add attack timeout and check if can reach
When about to attack an Entity :
Check if entity is withn reach
if not then cancel and set the target to null.

While chasing an entity :
If The dog stays in the same position for more than currently 2 seconds
he will cancel the attack.

Akso :
Make dog check for nearest Skeleton around
the PLAYER instead of himself because that is
more practical.