Redefine GuardMode
Dogs in this mode should make sure that the radius (GUARD_DISTANCE = 5)
around the owner is clear of enermy, if there
is enermy in the radius, the dog will attack it
until it either leaves the area or dies.
The mode also have 3 variant:
GUARD : Dog will attack every enermy
GUARD_FLAT : Same as GUARD but the dog won't
attack Piglins and Endermans
GUARD_MINOR : Dog won't do any attack but
growl to warn the owner if danger is
in the radius. The dog will
groul at Max Volume.