Added DogAsyncTaskManager System
This system is added for resolving requested "promises"
that would need to "tick" several time to resolve.

These promise works as follows :
When user want to request to execute a task that
require to be "ticked" over several tick to be resolved,
that will be submitted via addPromise() to the system.
The promise first being in a PENDING state when added.
In the next update every PENDING tick will get state
set to RUNNING, and that is when they start to tick
to resolve, when they failed or success to resolve they change
the state the REJECTED and FULFILLED respectively.
After that they do their final cleanUp to prevent any
resources leak and get removed from list.