Fix dog improper dismounting underwater
When a dog riding a player and the player
enters the water, the dog silently dismount serverside
but did not notify the client, this causes the client to still think
that the dog is not dismounted and still renders the dog on the player
head, thus the dog still has isOrderedToSit == true
serverside so even the client sees the dog as if he is
still riding and following the owner,  on serverside,
the dog was dismounted, stays and float
in the place that he silently dismounted serverside.
This causes a bug where the dog suddenly tp back to
"nowhere" when the player go somewhere else still having
the dog on head after dunking the dog in water, and then
dismount the dog then.