Redesign Dog GUI and some related systems
The old GUI however can still be accessible via config.
Created a mini UI framework which this ui is based on which took great
inspiration from Flux/Redux/ReactJS UI system. This allow
efficent codebase managing of the ui with ease.
In other words, future feature which require the gui to be updated
like perks or new and improved training system,
to be implemented easily in the gui with minimal
risk for bugs, And existing change decision to gui to be
carried out easier, more reliable, faster and more predictable.
Accessories now can be added and removed via the gui, which renders
the collar shear useless and ready for removal in the future
after another item being in design to asist in leveling down
talents and untaming or changing owner is added.
Stats Gui is now implemented, however its system still needs a major
redesign to decrease boilerplates, allow dynamic stats
entry (like each talents can have its own stats)
and managability.
There is now an option to disable the dog inventory button from
rendering in the inventory to avoid clashes with
some other mod that do something with that same spot
(ex: Quark).