Revamped skin system and dog rendering
Added a wrapper around skin payload representation
(which is solely its resourceloc) (not to be confused with
its reference from dog, which is a sha1 hash value of its texture file)
which includes model using, ear type and name of the texture.
The skins and its metadata is loaded when resource is being reloaded
by reading each resource_pack's 'skin.json'.
The server skin sync is and will still be present, but recieved skin
on client will have a default wrapper with default model and no name.
DogRenderer now can switch between models to render each skins.
Add offset constant in DogModel when the dog is lying or sitting.
added support for model which is using 64x64 texture size.