Overhaul Radar mechanic
Users will be presented with a screen
to pick dogs to locate, after they selected
the dog, the locating dog data is then being
written into the item nbt tag of the stack,
Right click again to clear.

If user is holding a Radar with the tag,
there will be a text rendered in the direction of where the dog is telling
user how far away.