Doggy Tools Talent ⛏️
This talent give dogs a toolbar where dog can make uses of specific
tools.
For example :
If a dog has a target, he will find the
nearest weapon from left to right of
the toolbar and equip.
If a dog has a hoe in the toolbar, the dog
will help you farming!