Imrpove Dog Navigation
+ MoveControl : Dogs now navigate with standard speed if
target y is lower than -0.75 and slower if lower than -1.75

+ PathNavigation : Properly center the target point, Minecraft did some
weird stuff here which cause the nextNode target center
to be offset by a fixed direction with the length depends on
the size of bbWidth. Properly cut corner, by defeault there
is corner cutting even when there is an questionable node at the
path after cutting corner.