Balancing out push prevention
+ Only avoid pushing each other when the other dog is navigating
or is not on ground. This causes less clipping, thus cause less
flickering and lowering the chance of causing a clipped group
to powerfully push you away.
+ Also disabling dog pushing any player in case a clipped group
does form (which is pretty unlikely) and also this also completely
solve the cases where dog pushing owner off cliffs or into some ?
place.