Extends dimension to include passangers as well
Dramatic improvement in wolf mount/mob
retriever. This redefines the dog's dimensions when there is a changes
in passanger leading to some improvements.
1. Pathfinding take the total dimension in to account and path find for
the total volume not just the dog volume
this avoid collision of the passanger when
the dog pathfind into narrow area
2. Collisons happens with the outer collison box, which collides
player when they are riding the dog
as well, prevents them from riding
the dog in narrow area and suffocate.