Duplication detection improvement
Each time a dogs is respawned from respawn storage, it will be granted
a unique sessionUUID(which is the default randomized uuid before setting
back to the old one) and will persist until the dog dies again.
The session uuid will be added to the dog's duplication detection tag when
saving to nbt.
A duplication is detected when the loading tag contains a uuid
which is asociated to the location storage but the session key is
different, signifying that dog is loaded some third-parties storages in
which the copy is saved before the the dog dies and go to respawn
storage.