Framework mechanic changes
Use a single root view for each StoreConnectedScreen
which holds all UI Content, and every
storeUpdate will be forward to it instead.
The screen dont have to worry about it.
The screen only need to reRender it when size changes.
init() would only be called once during the start of the lifetime
of the screen.