Imrpove dog ground navigation
Disable "Cutting Corner" functionailiy:
It proved not having significant affect on dog pathfinding in general.
If anything, it is the cause of a small chance of a dog tripping, especially small dogs.
Also use a static value for smaller dogs maxDistToWaypoint value
and tuned to cover most case (PIANO & PIANISSIMO)