fix a condition causing sittingPose remain even when not orderedToSit
When setOrderToSit to false while dog isInSittingPose and the sitGoal
hasn't started yet. This is because upon being read from save, sittingPose will be set to isOrderedToSit which is saved.
When a dog is joining the word, after the saved data being read, if the dog is
previously sitting,
the dog will be both isOrderedTOSit and isInSittingPose at the same time.
At this point no Goal is started yet. Usually the favorable case will be the SitGoal started after that and take ownership of isInSittingPose, setting it to false when stop ordered
Although Sometime the DTN Sit-override Action start before that setting isOrderedToSit to false stopping
The sitGoal from taking ownership of isInSittingPose, and since the sitGoal
is the only one in charge of setting it after read, the dog ended up in the state