Revamp Tactical Mode
The formal definition for Tactical Mode is the following: Your dog behaves like Agressive Mode but wil respond when being commanded it to attack something far away, currently via the Tactical Whistle
When the dog is in Tactical Mode targets a target far away:
+ It will focus on the target and any Aggressive mode behaviour that cause the Dog
to change target will not affect it. The target can, however, be changed with another
target when you issues another tactical command on another target or invalidated if
the player interupts the Dog.
+ Its MaxDistAwayFromOwner will be increased by 10 as long as it is still targeting
it.
+ Its follow range will be 32 instead of 16 until it gets close enough to the target.

Change to Tactical Whistle projectile:
+ They can now pass through allies.
+ (Configurable) There is now a limit on how many Tactical dogs will responds
each time it hit a target.
+ It now have a 0.5 second cooldown every use and every successful hit will have
an additional 2 second.