Dedicated Projectile Ally Check Handler for Dogs.
This use a head injection mixin into projectile's canHitEntity method
which only applies if either the owner of the proj or the target is a
Dog that prevent the projectile from seeing allies as a hit-able target
This allow piercing arrows to retain its percing value when skipping allies. And also allows Dog's Non DTN's Projectile to regconize allies and retain their piercing value if they are arrows as well.