(Config-able) Add corruption protection for dog who is riding player
In vanilla, if an entity is riding the player and Minecraft unexpectedly crashes without stopping properly, the entity will be corrupted since (1) Minecraft don't load passengers from players and the entity is saved as a passenger of that player instead of saving seperately when the crash happens.
This protection allow the dog to still be saved seperately while on the player's head, thus allowing the dog to persist when the crash happens. The dog (for now) will
still also be saved to the player's passengers tag but the tag will be omitted when load anyway because of (1) and we are also going to dismount the dog everytime the player about to log out.