DogAllyCheck improvement
We can now detect allied owner for other tamables whos owner is in a different dimension.
If the target is a tamable with a different owner who being offline in a level,
we try to locate the owner in the MinecraftServer instance
and the team the owner is on and call check against
that team instead.

We avoid calling owner.isAlliedTo during the allyCheck when target is a Dog to avoid trigger the mixin and unecessarily calling the allyCheck again at that point.

We only apply special tamable logic for tamed tamable.