Improve nav when dog about to approach "Danger Node"
A "Danger Node" in this context is defined as a path node with a danger
type and either it is the endNode of the path or its y is different
than the prevNode's y (or the dog's y block if it is the
current next node).
When there is a Danger Node no greater than 2 nodes away from the dog current next node in the dog current path, the dog will begin
sneaking, the closer the dog is, the slower the dog sneak.
Also other dog will not push that dog and vice versa.