GreedyFireSafeSearchPath fix
Only check the "corners" of the dog's bb for startNode if the blockpos that the dog is standing on is not valid.
if we need to start at a pos whose blockpos is not equals to the dog's blockpos, add it into the path, don't skip it.
If dog is not on ground (falling), try checking one block below if ground pos is air. This is needed because dog surfaceY is ground + 0.5.
Probably prioritize other pos over last_resort pos when scan surrounding for pos.